Rename `contrib` folder to `tensorflow`
