+++ docs/contributing.md
@@ -237,5 +237,5 @@ release!). Contributions are certainly not limited to these suggestions!
 
 -   **TF_Trajectories**. The source code currently includes a batch inference
     for running a batch of episodes using Tensorflow directly from C++ (in
-    `contrib/`). It has not yet been tested with CMake and public Tensorflow. We
+    `tensorflow/`). It has not yet been tested with CMake and public Tensorflow. We
     would like to officially support this and move it into the core library.
